chore: cargo clippy --fix auto-applied 671 warnings (1218→547) (v0.5.353)

+++ crates/chart/src/drawing/primitives_v2/fibonacci/trend_extension.rs
@@ -235,7 +235,7 @@ impl Primitive for FibTrendExtension {
             let in_bounds = if self.extend_right {
                 screen_x >= min_x
             } else {
-                screen_x >= min_x && screen_x <= viewport.chart_width as f64
+                screen_x >= min_x && screen_x <= viewport.chart_width
             };
 
             if in_bounds && (screen_y - level_y).abs() < HIT_TOLERANCE {
